requirements-win.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
pillow==10.1.0
opencv-python==4.8.1.78
insightface==0.7.3
faiss-cpu==1.7.4
scikit-learn==1.3.2
numpy==1.26.2
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.12.2
pandas==2.1.4
tqdm==4.66.1
psutil==5.9.6
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Предварительно собранные для Windows
# Используем wheel для dlib (если не установится, установите вручную)
dlib
face-recognition==1.3.0
face-recognition-models==0.3.0
